feat: :sparkles: add script execution functionality in main program mode
In this commit, the __name__ == "__main__" check was added to the end of the script, allowing the execution of the train function with parameters from the configuration file when the script is run as the main program. A message 'Training parameters were taken from the config file' has been added to indicate the process of loading training parameters. Also code to load the configuration file was added to use params from config in training